always use long long ll! ll! ll! ll!

+++ astro and bunch of words brute force.cpp
@@ -1,13 +1,13 @@
 #include<bits/stdc++.h>
-#define int long long
+#define ll long long
 #define ld long double
-#define pii pair<int,int>
+#define pii pair<ll,ll>
 #define fast_io  ios_base::sync_with_stdio(false);  cin.tie(NULL); cout.tie(NULL);
 #define peek(v) for(auto x:v) cout<<x<<" ";cout<<"\n";
-#define dpeek(v) for(vector<int> i : v) {for(int j : i){ cout<<j<<" ";} cout<<"\n";}
+#define dpeek(v) for(vector<ll> i : v) {for(ll j : i){ cout<<j<<" ";} cout<<"\n";}
 #define in_range(x, y, r, c) (x >= 0 && x < r && y >= 0 && y < c)
 using namespace std;
-void checkPermutations(int idx, vector<string> &p, int &ans){
+void checkPermutations(ll idx, vector<string> &p, ll &ans){
 	if(idx==p.size()){
 		return;
 	}
@@ -15,32 +15,32 @@ void checkPermutations(int idx, vector<string> &p, int &ans){
 	for(string &s : p){
 		toCheck.append(s);		
 	}
-	int i = 0; int j = toCheck.length()-1;
+	ll i = 0; ll j = toCheck.length()-1;
 	bool isPalindrome = true;
 	while(i<j){
 		if(toCheck[i]==toCheck[j]){i++;j--;}
 		else {isPalindrome = false; break;}
 	}
-	if(isPalindrome) ans = max(ans, (int)toCheck.length());
-	for(int i = idx+1; i<p.size(); i++){
+	if(isPalindrome) ans = max(ans, (ll)toCheck.length());
+	for(ll i = idx+1; i<p.size(); i++){
 		swap(p[idx], p[i]);
 		checkPermutations(idx+1,p,ans);
 		swap(p[idx], p[i]);
 	}
 }
-int maxConcatenatedPalindrome(vector<string>& words){
-	int ans = 0;
-	int n = words.size();
-	for(int i = (1ll<<n)-1; i>=0; i--){
+ll maxConcatenatedPalindrome(vector<string>& words){
+	ll ans = 0;
+	ll n = words.size();
+	for(ll i = (1ll<<n)-1; i>=0; i--){
 		vector<string> p;
-		for(int j = 0; j<n; j++){
+		for(ll j = 0; j<n; j++){
 			if(i&(1ll<<j)) p.push_back(words[j]);
 		}
 		checkPermutations(0,p,ans);
 	}
 	return ans;
 }
-int32_t main(){
+int main(){
 	fast_io;
 	vector<string> words1 = {"y", "xyx", "abc", "cba", "bac", "y", "y"};
     vector<string> words2 = {"y", "xyx", "abc", "cba", "bac", "y"};
